# api/core.openapi.yaml
basePath: /
definitions:
...
  CallerProfileResponse:
    properties:
      privileges:
        items:
          $ref: '#/definitions/PrivilegeKey'
        type: array
      roles:
        items:
          $ref: '#/definitions/CallerRoleGrant'
        type: array
      user:
        $ref: '#/definitions/User'
    type: object
  CallerRoleGrant:
    properties:
      granted_at:
        type: string
      privileges:
        items:
          $ref: '#/definitions/PrivilegeKey'
        type: array
        x-nullable: true
      role:
        $ref: '#/definitions/Role'
    required:
    - granted_at
    - role
    type: object
...
  PrivilegeKey:
    enum:
    - core:clusters:read
    - core:clusters:write
    - core:allocations:read
    - core:allocations:write
    - core:projects:read
    - core:projects:write
    - core:users:read
    - core:users:write
    - core:organizations:read
    - core:organizations:write
    - core:traces:read
    - core:privileges:grant
    - core:roles:manage
    - core:events:manage
    type: string
    x-enum-varnames:
    - ClustersRead
    - ClustersWrite
    - AllocationsRead
    - AllocationsWrite
    - ProjectsRead
    - ProjectsWrite
    - UsersRead
    - UsersWrite
    - OrganizationsRead
    - OrganizationsWrite
    - TracesRead
    - PrivilegesGrant
    - RolesManage
    - EventsManage
...
  Role:
    properties:
      created_at:
        type: string
      description:
        type: string
        x-nullable: true
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
    required:
    - created_at
    - id
    - is_system
    - name
    type: object
...
  User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      organization_id:
        type: string
      status:
        $ref: '#/definitions/UserStatus'
      type:
        $ref: '#/definitions/UserType'
    required:
    - email
    - first_name
    - id
    - last_name
    - organization_id
    - status
    - type
    type: object
...
  UserPrivilege:
    properties:
      granted_at:
        type: string
      granted_by:
        type: string
        x-nullable: true
      id:
        type: string
      privilege:
        $ref: '#/definitions/PrivilegeKey'
      reason:
        type: string
        x-nullable: true
      user_id:
        type: string
    required:
    - granted_at
    - id
    - privilege
    - user_id
    type: object
  UserRole:
    properties:
      granted_at:
        type: string
      granted_by:
        type: string
        x-nullable: true
      reason:
        type: string
        x-nullable: true
      role_id:
        type: string
      user_id:
        type: string
    required:
    - granted_at
    - role_id
    - user_id
    type: object
  UserStatus:
    enum:
    - PENDING
    - ACTIVE
    - INACTIVE
    - SUSPENDED
    - MERGED
    type: string
    x-enum-varnames:
    - UserPending
    - UserActive
    - UserInactive
    - UserSuspended
    - UserMerged
  UserType:
    enum:
    - CLUSTER_LOCAL
    - VIRTUAL
    - SYSTEM
    type: string
    x-enum-varnames:
    - UserTypeClusterLocal
    - UserTypeVirtual
    - UserTypeSystem
...
  grantPrivilegeRequest:
    properties:
      privilege:
        $ref: '#/definitions/PrivilegeKey'
      reason:
        type: string
    type: object
  grantRoleRequest:
    properties:
      reason:
        type: string
      role_id:
        type: string
    type: object
...
  revokePrivilegeRequest:
    properties:
      reason:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'REST API for Apache Custos: organizations, users, projects, compute
    clusters, allocations, audit traces, and the privilege/role authorization layer.
    OIDC bearer auth required. Connector endpoints live under /connectors/<name>/.'
  title: Apache Custos Core API
  version: 0.2.0
paths:
...
  /me:
    get:
      description: Roles are the caller's own grants; unlike /users/{id}/roles this
        needs no admin privilege.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CallerProfileResponse'
        "401":
          description: Missing authenticated caller
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get the caller's user record, effective privileges, and roles
      tags:
      - Caller
...
  /privileges/{key}/holders:
    get:
      description: Role-derived holders are NOT listed here.
      parameters:
      - description: Privilege key
        enum:
        - core:clusters:read
        - core:clusters:write
        - core:allocations:read
        - core:allocations:write
        - core:projects:read
        - core:projects:write
        - core:users:read
        - core:users:write
        - core:organizations:read
        - core:organizations:write
        - core:traces:read
        - core:privileges:grant
        - core:roles:manage
        - core:events:manage
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserPrivilege'
            type: array
        "400":
          description: Unknown privilege key
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List direct holders of a privilege
      tags:
      - Privileges
  /privileges/catalog:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PrivilegeKey'
            type: array
        "401":
          description: Missing or invalid bearer token
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Caller lacks privileges:grant
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List the declared privilege catalog
      tags:
      - Privileges
...
  /user/privileges:
    get:
      description: Effective set is direct grants UNION every privilege carried by
        every role the caller holds.
      produces:
      - application/json
      responses:
        "200":
          description: Effective privilege set
          schema:
            properties:
              privileges:
                items:
                  $ref: '#/definitions/PrivilegeKey'
                type: array
            type: object
        "401":
          description: Missing authenticated caller
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get caller's effective privileges
      tags:
      - Caller
...
  /users/{id}/privileges:
    get:
      description: Returns DIRECT grants only (not role-derived). Combine with `GET
        /users/{id}/roles` for the full picture.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserPrivilege'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Caller lacks privileges:grant
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List a user's direct privilege grants
      tags:
      - Privileges
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Grant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/grantPrivilegeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/UserPrivilege'
        "400":
          description: Bad request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Missing or invalid bearer token
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Caller lacks privileges:grant
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Privilege already active for user
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Grant a direct privilege to a user
      tags:
      - Privileges
  /users/{id}/privileges/{key}:
    delete:
      consumes:
      - application/json
      description: '`privileges:grant` itself cannot be self-revoked or revoked from
        the last holder.'
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Privilege key
        enum:
        - core:clusters:read
        - core:clusters:write
        - core:allocations:read
        - core:allocations:write
        - core:projects:read
        - core:projects:write
        - core:users:read
        - core:users:write
        - core:organizations:read
        - core:organizations:write
        - core:traces:read
        - core:privileges:grant
        - core:roles:manage
        - core:events:manage
        in: path
        name: key
        required: true
        type: string
      - description: Optional reason
        in: body
        name: request
        schema:
          $ref: '#/definitions/revokePrivilegeRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Self-revoke meta, last-holder, or unknown key
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Missing or invalid bearer token
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Caller lacks privileges:grant
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: No active grant for that key
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke a direct privilege from a user
      tags:
      - Privileges
  /users/{id}/roles:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserRole'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List roles a user holds
      tags:
      - Role Assignments
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/grantRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/UserRole'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: User already holds that role
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Grant a role to a user
      tags:
      - Role Assignments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"