# pkg/models/role.go
// Licensed to the Apache Software Foundation (ASF) under one
// or more contributor license agreements.  See the NOTICE file
// distributed with this work for additional information
// regarding copyright ownership.  The ASF licenses this file
// to you under the Apache License, Version 2.0 (the
// "License"); you may not use this file except in compliance
// with the License.  You may obtain a copy of the License at
//
//   http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing,
// software distributed under the License is distributed on an
// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
// KIND, either express or implied.  See the License for the
// specific language governing permissions and limitations
// under the License.

package models

import "time"

// SystemRoleSuperAdmin is the bootstrap role: carries privileges:grant
// and roles:manage. Created on first server start.
const SystemRoleSuperAdmin = "super_admin"

// SystemRoleAdmin carries every registered privilege except privileges:grant
// and roles:manage, so only super admins can change grants and roles.
const SystemRoleAdmin = "admin"

// Role is a named bundle of privileges. Mutating the bundle propagates to
// every holder.
type Role struct {
	ID          string    `json:"id"          db:"id" binding:"required"`
	Name        string    `json:"name"        db:"name" binding:"required"`
	Description *string   `json:"description" db:"description" extensions:"x-nullable"`
	IsSystem    bool      `json:"is_system"   db:"is_system" binding:"required"`
	CreatedAt   time.Time `json:"created_at"  db:"created_at" binding:"required"`
}

// UserRole is one role assignment. Revoke is DELETE; history lives in
// audit_events.
type UserRole struct {
	UserID    string    `json:"user_id"    db:"user_id" binding:"required"`
	RoleID    string    `json:"role_id"    db:"role_id" binding:"required"`
	GrantedBy *string   `json:"granted_by" db:"granted_by" extensions:"x-nullable"`
	GrantedAt time.Time `json:"granted_at" db:"granted_at" binding:"required"`
	Reason    *string   `json:"reason"     db:"reason" extensions:"x-nullable"`
}